+++ newsapp/static/newsapp/js/script.js
