+++ newsapp/templates/newsapp/register.html
