+++ newsapp/utils.py
