+++ newsapp/views.py
@@ -1,17 +1,13 @@
 from django.contrib.auth.views import LoginView
-from django.http import HttpResponse, HttpResponseNotFound, Http404
+from django.http import Http404
 from django.shortcuts import render
-from django.contrib.auth.models import User
-from django.contrib.auth import login, logout
+from django.contrib.auth import logout
 from django.shortcuts import redirect
 from django.urls import reverse_lazy
-from django.views.generic import ListView, DetailView, CreateView
-from django.contrib.auth.mixins import LoginRequiredMixin
+from django.views.generic import CreateView
 from .forms import FeedbackForm, RegisterUserForm, LoginUserForm, CommentsForm
-from newsapp.models import News, Category, Feedback, Comments
+from newsapp.models import News, Category, Comments
 from .utils import DataMixin
-
-categories = Category.objects.all()
 
 footer_menu = [
     {'title': 'Контакты', 'path_name': 'contacts'},
@@ -25,7 +21,7 @@ def index(request):
     context = {
         'title': 'NEWSRAGE',
         'posts': posts,
-        'categories': categories,
+        'categories': Category.objects.all(),
         'footer_menu': footer_menu,
     }
 
@@ -33,9 +29,9 @@ def index(request):
 
 
 def category(request, category_slug):
-    category = Category.objects.get(slug=category_slug)
+    page_category = Category.objects.get(slug=category_slug)
 
-    posts = News.objects.filter(category=category.pk).filter(is_published=True)
+    posts = News.objects.filter(category=page_category.pk).filter(is_published=True)
 
     if len(posts) == 0:
         raise Http404()
@@ -43,7 +39,7 @@ def category(request, category_slug):
     context = {
         'title': 'NEWSRAGE',
         'posts': posts,
-        'categories': categories,
+        'categories': Category.objects.all(),
         'footer_menu': footer_menu,
     }
 
@@ -71,7 +67,7 @@ def post(request, post_slug):
     context = {
         'title': 'NEWSRAGE',
         'post': post,
-        'categories': categories,
+        'categories': Category.objects.all(),
         'footer_menu': footer_menu,
         'post_comments': post_comments,
         'form': form,
@@ -87,7 +83,7 @@ def contacts(request):
     context = {
         'title': 'NEWSRAGE',
         'post': post,
-        'categories': categories,
+        'categories': Category.objects.all(),
         'footer_menu': footer_menu,
     }
     return render(request, 'newsapp/contacts.html', context=context)
@@ -108,7 +104,7 @@ def feedback(request):
     context = {
         'title': 'NEWSRAGE',
         'post': post,
-        'categories': categories,
+        'categories': Category.objects.all(),
         'footer_menu': footer_menu,
         'form': form,
     }
@@ -147,16 +143,3 @@ def logout_user(request):
 
 def pageNotFound(response, exception):
     return render(response, 'newsapp/notfound.html')
-
-# class Index(DataMixin, ListView):
-#     model = News
-#     template_name = 'newsapp/index.html'
-#     context_object_name = 'news'
-#
-#     def get_context_data(self, *, object_list=None, **kwargs):
-#         context = super().get_context_data(**kwargs)
-#         user_context = self.get_user_context()
-#         return dict(list(context.items()) + list(user_context.items()))
-#
-#     def get_queryset(self):
-#         return News.objects.filter(is_published=True)
