--- newsrage/docker-compose.yml
@@ -1,18 +1,0 @@
-version: "3.8"
-
-services:
-  db:
-    image: nouchka/sqlite3:latest
-    volumes:
-      - ./data/db:/root/db
-    environment:
-      - SQLITE3_DB=db.sqlite3
-  web:
-    build: .
-    command: python manage.py runserver 0.0.0.0:8000
-    volumes:
-      - .:/code
-    ports:
-      - "8000:8000"
-    depends_on:
-      - db
